--- backend/src/main/java/com/concordia/velocity/model/Operator.java
@@ -1,1 +1,0 @@
-package com.concordia.velocity.model;
