+++ backend/src/main/java/com/concordia/velocity/strategy/OneTimeStandardPayment.java
@@ -39,10 +39,6 @@ public class OneTimeStandardPayment implements PaymentStrategy {
 
 
 
-    @Override
-    public String getStrategyName() {
-        return "One-Time Standard Payment";
-    }
 
     /**
      * Gets the base price for this strategy
